requirements.txt
scipy>=0.18.1
numpy>=1.11.3
pytest>=3.0
scikit-learn>=0.18.1
matplotlib>=2.0.0
progressbar2>=3.16.0
biopython 
